chore: Update shuffleDrawStack to shuffle top card as well
This commit modifies the `shuffleDrawStack` function in the `GameManager` to include an additional parameter `shuffleTopCardToo`. When `shuffleTopCardToo` is set to `true`, the function shuffles the top card along with the rest of the cards in the draw stack. This change ensures that the top card is also randomized when shuffling the draw stack.