chore: Update SetStackStream function in GameManager
This commit modifies the `SetStackStream` function in the `GameManager` to include an additional parameter `initMsg *pb.StackRequest`. This parameter is used to send an initialization message to the stack stream when `initMsg.DrawingCard` is false. The commit also removes the commented out code for handling card draw messages, which is marked as a TODO.